Move checksum configs to aws_smithy_types
Update codegen to point to new checksum config
