+++ aws/rust-runtime/aws-config/external-types.toml
@@ -20,8 +20,6 @@ allowed_external_types = [
    "aws_smithy_async::rt::sleep::SharedAsyncSleep",
    "aws_smithy_async::time::SharedTimeSource",
    "aws_smithy_async::time::TimeSource",
-   "aws_smithy_checksums::RequestChecksumCalculation",
-   "aws_smithy_checksums::ResponseChecksumValidation",
    "aws_smithy_runtime::client::identity::cache::IdentityCache",
    "aws_smithy_runtime::client::identity::cache::lazy::LazyCacheBuilder",
    "aws_smithy_runtime_api::box_error::BoxError",
@@ -36,6 +34,8 @@ allowed_external_types = [
    "aws_smithy_runtime_api::client::result::SdkError",
    "aws_smithy_runtime_api::client::stalled_stream_protection::StalledStreamProtectionConfig",
    "aws_smithy_types::body::SdkBody",
+   "aws_smithy_types::checksum_config::RequestChecksumCalculation",
+   "aws_smithy_types::checksum_config::ResponseChecksumValidation",
    "aws_smithy_types::retry",
    "aws_smithy_types::retry::*",
    "aws_smithy_types::timeout",
