+++ aws/rust-runtime/aws-types/external-types.toml
@@ -6,8 +6,6 @@ allowed_external_types = [
     "aws_smithy_async::rt::sleep::SharedAsyncSleep",
     "aws_smithy_async::time::SharedTimeSource",
     "aws_smithy_async::time::TimeSource",
-    "aws_smithy_checksums::ResponseChecksumValidation",
-    "aws_smithy_checksums::RequestChecksumCalculation",
     "aws_smithy_runtime_api::client::behavior_version::BehaviorVersion",
     "aws_smithy_runtime_api::client::http::HttpClient",
     "aws_smithy_runtime_api::client::http::SharedHttpClient",
@@ -15,6 +13,8 @@ allowed_external_types = [
     "aws_smithy_runtime_api::client::identity::SharedIdentityCache",
     "aws_smithy_runtime_api::client::stalled_stream_protection::StalledStreamProtectionConfig",
     "aws_smithy_runtime_api::http::headers::Headers",
+    "aws_smithy_types::checksum_config::ResponseChecksumValidation",
+    "aws_smithy_types::checksum_config::RequestChecksumCalculation",
     "aws_smithy_types::config_bag::storable::Storable",
     "aws_smithy_types::config_bag::storable::StoreReplace",
     "aws_smithy_types::config_bag::storable::Storer",
