+++ aws/sdk-codegen/src/main/kotlin/software/amazon/smithy/rustsdk/HttpRequestChecksumDecorator.kt
@@ -220,8 +220,8 @@ class HttpRequestChecksumConfigCustomization(private val codegenContext: ClientC
             *preludeScope,
             "RequestChecksumCalculation" to
                 configReexport(
-                    RuntimeType.smithyChecksums(rc)
-                        .resolve("RequestChecksumCalculation"),
+                    RuntimeType.smithyTypes(rc)
+                        .resolve("checksum_config::RequestChecksumCalculation"),
                 ),
         )
