+++ aws/sdk-codegen/src/main/kotlin/software/amazon/smithy/rustsdk/HttpResponseChecksumDecorator.kt
@@ -191,8 +191,8 @@ class HttpResponseChecksumConfigCustomization(private val codegenContext: Client
             *preludeScope,
             "ResponseChecksumValidation" to
                 configReexport(
-                    RuntimeType.smithyChecksums(rc)
-                        .resolve("ResponseChecksumValidation"),
+                    RuntimeType.smithyTypes(rc)
+                        .resolve("checksum_config::ResponseChecksumValidation"),
                 ),
         )
