+++ rust-runtime/aws-smithy-checksums/external-types.toml
@@ -1,7 +1,5 @@
 allowed_external_types = [
     "aws_smithy_types::body::SdkBody",
-    "aws_smithy_types::config_bag::storable::StoreReplace",
-    "aws_smithy_types::config_bag::storable::Storable",
     "bytes::bytes::Bytes",
     "http::header::map::HeaderMap",
     "http::header::name::HeaderName",
