+++ rust-runtime/aws-smithy-types/src/lib.rs
@@ -20,6 +20,7 @@
 pub mod base64;
 pub mod body;
 pub mod byte_stream;
+pub mod checksum_config;
 /// A typemap for storing configuration.
 pub mod config_bag;
 pub mod date_time;
